@@ -1,9 +1,15 @@
 module github.com/NdoleStudio/lfu-cache
 
-go 1.13
+go 1.18
 
 require (
 	github.com/dgrijalva/lfu-go v0.0.0-20141010002404-f174e76c5138
 	github.com/pkg/errors v0.9.1
 	github.com/stretchr/testify v1.5.1
+)
+
+require (
+	github.com/davecgh/go-spew v1.1.0 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	gopkg.in/yaml.v2 v2.2.2 // indirect
 )
