@@ -10,18 +10,18 @@ import (
 
 func TestNew(t *testing.T) {
 	t.Run("it returns a pointer to the cache when the cap is valid", func(t *testing.T) {
-		cache, err := New(2)
+		cache, err := New[int, string](2)
 		assert.NoError(t, err)
-		assert.IsType(t, &Cache{}, cache)
+		assert.IsType(t, &Cache[int, string]{}, cache)
 	})
 
 	t.Run("it returns ErrInvalidCap when the cap is 0", func(t *testing.T) {
-		_, err := New(0)
+		_, err := New[int, string](0)
 		assert.EqualError(t, err, ErrInvalidCap.Error())
 	})
 
 	t.Run("it returns ErrInvalidCap when the cap is less than 0", func(t *testing.T) {
-		_, err := New(-2)
+		_, err := New[int, string](-2)
 		assert.Error(t, err)
 		assert.EqualError(t, err, ErrInvalidCap.Error())
 	})
@@ -29,14 +29,14 @@ func TestNew(t *testing.T) {
 
 func TestCache_Len(t *testing.T) {
 	t.Run("it returns the 0 when there are no items in the cache", func(t *testing.T) {
-		cache, err := New(4)
+		cache, err := New[int, string](4)
 
 		assert.NoError(t, err)
 		assert.Equal(t, 0, cache.Len())
 	})
 
 	t.Run("it returns the 3 when there are 3 items in the cache", func(t *testing.T) {
-		cache, err := New(4)
+		cache, err := New[string, int](4)
 		assert.NoError(t, err)
 
 		err = cache.Set("how", 1)
@@ -50,7 +50,7 @@ func TestCache_Len(t *testing.T) {
 	})
 
 	t.Run("it returns the size of the cache after more than cap items have been inserted", func(t *testing.T) {
-		cache, err := New(2)
+		cache, err := New[string, int](2)
 		assert.NoError(t, err)
 
 		err = cache.Set("how", 1)
@@ -66,13 +66,13 @@ func TestCache_Len(t *testing.T) {
 
 func TestCache_Cap(t *testing.T) {
 	t.Run("it returns the size of the cache after creation", func(t *testing.T) {
-		cache, err := New(34)
+		cache, err := New[string, int](34)
 		assert.NoError(t, err)
 		assert.Equal(t, 34, cache.Cap())
 	})
 
 	t.Run("it returns the size of the cache after items have been added in the cache", func(t *testing.T) {
-		cache, err := New(2)
+		cache, err := New[string, int](2)
 		assert.NoError(t, err)
 
 		err = cache.Set("you", 1)
@@ -82,7 +82,7 @@ func TestCache_Cap(t *testing.T) {
 	})
 
 	t.Run("it returns the size of the cache after more than cap items have been inserted", func(t *testing.T) {
-		cache, err := New(2)
+		cache, err := New[string, int](2)
 		assert.NoError(t, err)
 
 		err = cache.Set("how", 1)
@@ -98,14 +98,14 @@ func TestCache_Cap(t *testing.T) {
 
 func TestCache_IsFull(t *testing.T) {
 	t.Run("it returns false after cache is initialised", func(t *testing.T) {
-		cache, err := New(2)
+		cache, err := New[string, int](2)
 		assert.NoError(t, err)
 
 		assert.False(t, cache.IsFull())
 	})
 
 	t.Run("it returns true when the cache is full", func(t *testing.T) {
-		cache, err := New(2)
+		cache, err := New[string, int](2)
 		assert.NoError(t, err)
 
 		err = cache.Set("how", 1)
@@ -117,7 +117,7 @@ func TestCache_IsFull(t *testing.T) {
 	})
 
 	t.Run("it returns true after more than cap items have been inserted", func(t *testing.T) {
-		cache, err := New(2)
+		cache, err := New[string, int](2)
 		assert.NoError(t, err)
 
 		err = cache.Set("how", 1)
@@ -133,14 +133,14 @@ func TestCache_IsFull(t *testing.T) {
 
 func TestCache_IsEmpty(t *testing.T) {
 	t.Run("it returns true after cache is initialised", func(t *testing.T) {
-		cache, err := New(2)
+		cache, err := New[string, int](2)
 		assert.NoError(t, err)
 
 		assert.True(t, cache.IsEmpty())
 	})
 
 	t.Run("int returns false when there are items in the cache", func(t *testing.T) {
-		cache, err := New(2)
+		cache, err := New[string, int](2)
 		assert.NoError(t, err)
 
 		err = cache.Set("how", 1)
@@ -150,7 +150,7 @@ func TestCache_IsEmpty(t *testing.T) {
 	})
 
 	t.Run("int returns false when the cache is full", func(t *testing.T) {
-		cache, err := New(2)
+		cache, err := New[string, int](2)
 		assert.NoError(t, err)
 
 		err = cache.Set("how", 1)
@@ -163,7 +163,7 @@ func TestCache_IsEmpty(t *testing.T) {
 	})
 
 	t.Run("it returns false after more than cap items have been inserted", func(t *testing.T) {
-		cache, err := New(2)
+		cache, err := New[string, int](2)
 		assert.NoError(t, err)
 
 		err = cache.Set("how", 1)
@@ -179,7 +179,7 @@ func TestCache_IsEmpty(t *testing.T) {
 
 func TestCache_Set(t *testing.T) {
 	t.Run("it returns ErrInvalidCap when the cap rate is 0", func(t *testing.T) {
-		cache := &Cache{}
+		cache := &Cache[string, int]{}
 
 		err := cache.Set("how", 1)
 
@@ -187,7 +187,7 @@ func TestCache_Set(t *testing.T) {
 	})
 
 	t.Run("it returns ErrInvalidCap when the cap rate is < 0", func(t *testing.T) {
-		cache := &Cache{cap: -1}
+		cache := &Cache[string, int]{cap: -1}
 
 		err := cache.Set("how", 1)
 
@@ -195,7 +195,7 @@ func TestCache_Set(t *testing.T) {
 	})
 
 	t.Run("it can set a nil value in the cache", func(t *testing.T) {
-		cache, err := New(1)
+		cache, err := New[string, *int](1)
 		assert.NoError(t, err)
 
 		err = cache.Set("how", nil)
@@ -203,7 +203,7 @@ func TestCache_Set(t *testing.T) {
 	})
 
 	t.Run("it doesn't add a new value if the key is already in the cache", func(t *testing.T) {
-		cache, err := New(2)
+		cache, err := New[string, int](2)
 		assert.NoError(t, err)
 
 		key := "how"
@@ -221,7 +221,7 @@ func TestCache_Set(t *testing.T) {
 	})
 
 	t.Run("it can insert more than cap values in the cache", func(t *testing.T) {
-		cache, err := New(2)
+		cache, err := New[string, int](2)
 		assert.NoError(t, err)
 
 		err = cache.Set("how", 1)
@@ -236,7 +236,7 @@ func TestCache_Set(t *testing.T) {
 	})
 
 	t.Run("it removes the LFU item in the cache when inserting above capacity", func(t *testing.T) {
-		cache, err := New(2)
+		cache, err := New[string, int](2)
 		assert.NoError(t, err)
 
 		err = cache.Set("how", 1)
@@ -267,7 +267,7 @@ func TestCache_Set(t *testing.T) {
 
 func TestCache_Get(t *testing.T) {
 	t.Run("it returns ErrCacheMiss if an item is not in the cache", func(t *testing.T) {
-		cache, err := New(2)
+		cache, err := New[string, int](2)
 		assert.NoError(t, err)
 
 		_, err = cache.Get("are")
@@ -275,7 +275,7 @@ func TestCache_Get(t *testing.T) {
 	})
 
 	t.Run("it returns an item when it is in the cache", func(t *testing.T) {
-		cache, err := New(2)
+		cache, err := New[string, int](2)
 		assert.NoError(t, err)
 
 		err = cache.Set("are", 2)
@@ -287,21 +287,24 @@ func TestCache_Get(t *testing.T) {
 	})
 
 	t.Run("it can get a nil value in the cache", func(t *testing.T) {
-		cache, err := New(1)
+		cache, err := New[string, *int](1)
 		assert.NoError(t, err)
 
-		err = cache.Set("how", nil)
+		var value *int = nil
+
+		err = cache.Set("how", value)
 		assert.NoError(t, err)
 
 		val, err := cache.Get("how")
 		assert.NoError(t, err)
-		assert.Equal(t, nil, val)
+
+		assert.Equal(t, value, val)
 	})
 }
 
 func TestInternalImplementation(t *testing.T) {
 	t.Run("the internal storage items length is equal to 1 when there is 1 item in the cache", func(t *testing.T) {
-		cache, err := New(3)
+		cache, err := New[string, int](3)
 		assert.NoError(t, err)
 
 		key := "how"
@@ -339,7 +342,7 @@ func TestInternalImplementation(t *testing.T) {
 	// 2 => "you" -> "today" -> nil
 	// 3 => "how" -> nil
 	t.Run("the internal storage items length when inserting above capacity", func(t *testing.T) {
-		cache, err := New(4)
+		cache, err := New[string, int](4)
 		assert.NoError(t, err)
 
 		err = cache.Set("how", 1)
@@ -383,7 +386,7 @@ func TestInternalImplementation(t *testing.T) {
 }
 
 func BenchmarkCache(b *testing.B) {
-	cache, err := New(100)
+	cache, err := New[string, int](100)
 	if err != nil {
 		b.Fatal(err.Error())
 	}
